Improve types for bitset

+++ bitset.go
@@ -7,14 +7,14 @@ package schulze
 
 type bitSet []uint64
 
-func newBitset(size uint) bitSet {
+func newBitset(size uint64) bitSet {
 	return bitSet(make([]uint64, size/64+1))
 }
 
-func (s bitSet) set(i uint) {
+func (s bitSet) set(i uint64) {
 	s[i/64] |= 1 << (i % 64)
 }
 
-func (s bitSet) isSet(i uint) bool {
+func (s bitSet) isSet(i uint64) bool {
 	return s[i/64]&(1<<(i%64)) != 0
 }
